Serving/requirements.txt
pyonmttok>=1.10.1,<2
tensorflow-serving-api==1.11.0
pymongo==3.7.2
